requirements.txt
numpy==1.23.5
pytorch-lightning==1.9.3
torchvision==0.14.1
torchaudio==0.13.1
gdown==4.6.4
progressbar==2.5
tqdm==4.64.1
jupyterlab==3.5.3
numba==0.56.4
pyloudnorm==0.1.1
pandas==1.5.2
pydub==0.25.1
hydra-core==1.3.2
torch-audiomentations==0.11.0
seaborn
librosa
matplotlib
pytest